requirements.txt
fastapi
uvicorn[standard]
pillow

